#10822 Schema/StatsBundle: Fixed viewsLog events.
