+++ src/Pumukit/StatsBundle/Resources/config/services.xml
@@ -4,9 +4,8 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
-    
     <services>
-        <service id="pumukit_stats.log" class="Pumukit\StatsBundle\Services\LogService">
+        <service id="pumukit_stats.log" class="Pumukit\StatsBundle\EventListener\Log">
             <argument type="service" id="doctrine_mongodb.odm.document_manager" />
             <argument type="service" id="request_stack" />
             <argument type="service" id="security.token_storage"/>
@@ -15,9 +14,7 @@
 
         <service id="pumukit_stats.stats" class="Pumukit\StatsBundle\Services\StatsService">
             <argument type="service" id="doctrine_mongodb.odm.document_manager" />
-        </service>        
+        </service>
     </services>
 
-    
-   
 </container>
